+++ CHANGELOG.md
@@ -38,6 +38,7 @@ is based on https://keepachangelog.com/en/1.1.0/.
   `packet::payload::InstructionTrace::as_support` and
   `packet::payload::Payload::as_support` for accessing the
   `packet::sync::Support` they may hold.
+- `Display` impl for `instruction::bits::Bits`.
 
 ### Changed